+++ isis3/galileo_iss/processing.py
@@ -41,7 +41,7 @@ def is_supported_file(file_name):
     else:
         return False
 
-def process_pds_data_file(from_file_name, is_ringplane=False, is_verbose=False, skip_if_cub_exists=False, init_spice=True, **args):
+def process_pds_data_file(from_file_name, is_verbose=False, skip_if_cub_exists=False, init_spice=True, nocleanup=False, additional_options={}):
     product_id = info.get_product_id(from_file_name)
 
     out_file_tiff = "%s.tif" % output_filename(from_file_name)
@@ -50,6 +50,11 @@ def process_pds_data_file(from_file_name, is_ringplane=False, is_verbose=False, 
     if skip_if_cub_exists and os.path.exists(out_file_cub):
         print "File %s exists, skipping processing" % out_file_cub
         return
+
+    if "ringplane" in additional_options:
+        is_ringplane = additional_options["ringplane"].upper() in ("TRUE", "YES")
+    else:
+        is_ringplane = False
 
     source_dirname = os.path.dirname(from_file_name)
     if source_dirname == "":
@@ -121,11 +126,17 @@ def process_pds_data_file(from_file_name, is_ringplane=False, is_verbose=False, 
     if is_verbose:
         print s
 
-    if is_verbose:
-        print "Cleaning up..."
+    if nocleanup is False:
+        if is_verbose:
+            print "Cleaning up..."
+        else:
+            printProgress(7, 9, prefix="%s: "%from_file_name)
+        map(os.unlink, glob.glob('%s/__%s*.cub'%(work_dir, product_id)))
     else:
-        printProgress(7, 9, prefix="%s: "%from_file_name)
-    map(os.unlink, glob.glob('%s/__%s*.cub'%(work_dir, product_id)))
+        if is_verbose:
+            print "Skipping clean up..."
+        else:
+            printProgress(7, 9, prefix="%s: "%from_file_name)
 
     if not is_verbose:
         printProgress(9, 9, prefix="%s: "%from_file_name)
